Refactor to use Vitejs.
* Update/simplified Prettier setup.
