--- .editorconfig
@@ -1,16 +1,0 @@
-# http://editorconfig.org
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*.{py,html}]
-indent_size = 4
-
-[*.md]
-trim_trailing_whitespace = false
