+++ .eslintrc.js
@@ -1,0 +1,19 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+  env: {
+    browser: true,
+    es2021: true,
+  },
+  extends: ['eslint:recommended', 'plugin:react/recommended', 'plugin:prettier/recommended'],
+  plugins: ['prettier'],
+  parserOptions: {
+    ecmaVersion: 12,
+    sourceType: 'module',
+  },
+  rules: {
+    'no-undef': 'off',
+    'prettier/prettier': 'error',
+    'react/prop-types': 0,
+    'react/no-unescaped-entities': 0,
+  },
+}
