+++ .prettierrc
@@ -1,7 +1,6 @@
 {
-	"semi": false,
-	"singleQuote": true,
-	"tabs": true,
-	"useTabs": true,
-	"trailingComma": "es5"
+  "printWidth": 80,
+  "semi": false,
+  "singleQuote": true,
+  "tabWidth": 2
 }
