+++ postcss.config.js
@@ -1,9 +1,3 @@
 module.exports = {
-	plugins: [
-		require(`postcss-import`),
-		require(`tailwindcss/nesting`),
-		require(`tailwindcss`),
-		require('postcss-custom-properties'),
-		require(`autoprefixer`),
-	],
+  plugins: [require(`postcss-import`), require(`tailwindcss/nesting`), require(`tailwindcss`), require(`autoprefixer`)],
 }
