--- public/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width,initial-scale=1" />
-		<title>Holly Template</title>
-		<link
-			rel="stylesheet"
-			href="https://fonts.googleapis.com/css?family=Fira+Sans:400,400i,500,700"
-		/>
-		<link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-		<link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-	</head>
-
-	<body class="bg-gray-100">
-		<noscript> You need to enable JavaScript to run this app. </noscript>
-		<div id="app"></div>
-	</body>
-</html>
