--- src/components/background-gradient.js
@@ -1,12 +1,0 @@
-import classnames from 'clsx'
-
-export function BackgroundGradient({ className }) {
-	return (
-		<div
-			className={classnames(
-				'bg-gradient-to-b from-primary-600 to-primary-400',
-				className
-			)}
-		/>
-	)
-}
