+++ src/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+
+import './assets/styles.css'
+import App from './app'
+
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+)
