+++ vite.config.js
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  root: 'src',
+  base: '/holly-react/',
+  build: {
+    outDir: '../dist',
+  },
+  plugins: [react()],
+})
